Сделан функционал добавления магазина. Добавлено приложение api_domer. Исправлены урлы личного кабинета в urls.py и шаблонах
