+++ api_domer/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class ApiDomerConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'api_domer'
