+++ main_page_domer/static/css/my_store.css
@@ -1,0 +1,5 @@
+.thin_line {
+    margin-bottom: 5px;
+    height: 1px;
+    background: #D4D4D4;
+}
