+++ main_page_domer/static/js/add_store.js
@@ -6,7 +6,7 @@ let region_hidden = document.querySelector(".region_hidden")
 
 function show_city(event){
     if (event.target.value !== "0"){
-        fetch(`http://127.0.0.1:8000/add_store/city/${event.target.value}`)
+        fetch(`http://127.0.0.1:8000/api/v1/add_store/city/${event.target.value}`)
             .then((response) => response.json())
             .then(data => {
                 region_hidden.style.display = "flex";
@@ -22,7 +22,7 @@ function show_city(event){
 
 let get_category = document.querySelector("#select_category");
 
-fetch("http://127.0.0.1:8000/add_store/categories/")
+fetch("http://127.0.0.1:8000/api/v1/add_store/categories/")
     .then((response) => response.json())
     .then((data) => {
         get_category.innerHTML += data.map((category) => {
