+++ main_page_domer/templates/add_store.html
@@ -1,7 +1,7 @@
 {% extends 'base.html' %}
 
 {% load static %}
-{% load mptt_tags %}
+
 {% block content %}
     <title>Добавление магазина</title>
     <div class="thin_line"></div>
