+++ main_page_domer/templates/base_for_personal_account.html
@@ -7,6 +7,7 @@
     <link rel="stylesheet" href="{% static 'css/personal_account_style.css' %}">
     <link rel="stylesheet" href="{% static 'css/user_data.css' %}">
     <link rel="stylesheet" href="{% static 'css/messages.css' %}">
+    <link rel="stylesheet" href="{% static 'css/my_store.css' %}">
 
 </head>
 <body>
