+++ main_page_domer/urls.py
@@ -12,6 +12,5 @@ urlpatterns = [
     path('sent_messages/', get_sent_page, name='sent_messages'),
     path('admin_messages/', get_admin_message_page, name='admin_messages'),
     path('add_store/', add_store, name='add_store'),
-    path('add_store/city/<int:id>', get_list_of_cities, name='list_of_cities'),
-    path('add_store/categories/', get_list_of_categories, name='list_of_categories'),
+    path('my_store/', get_my_store, name='my_store')
 ]
